% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{ensure_level_is_auditory_level}
\alias{ensure_level_is_auditory_level}
\title{Ensure that an object's level is an auditory level}
\usage{
ensure_level_is_auditory_level(object)
}
\description{
Ensure that an object's level is an auditory level
}
\examples{
ensure_level_is_auditory_level(
  new("pure_tone", frequency = as.kHz(0.44), level = as.dB_SPL(80))
)
}
