% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{get_audibility}
\alias{get_audibility}
\title{Get audibility}
\usage{
get_audibility(audible_level, al_0 = 15)
}
\arguments{
\item{audible_level}{Numeric vector of audible levels (dB)}

\item{al_0}{constant (see Equation 7 of Parncutt & Strasburger (1994).)}
}
\description{
Returns the audibility of a set of pure tone components as a function of their audible levels. Corresponds to Equation 7 of Parncutt & Strasburger (1994).
}
