% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{get_combined_spectrum}
\alias{get_combined_spectrum}
\title{Get combined spectrum}
\usage{
get_combined_spectrum(pure_midi_pitch, pure_tone_audibility, complex_midi_pitch,
  complex_tone_audibility, k_s)
}
\arguments{
\item{pure_midi_pitch}{Numeric vector of MIDI pitches for the pure spectrum}

\item{pure_tone_audibility}{Numeric vector of audibilities for the pure spectrum}

\item{k_s}{Numeric scalar; parameter from Parncutt & Strasburger (1994)}

\item{pure_midi_pitch}{Numeric vector of MIDI pitches for the complex spectrum}

\item{pure_tone_audibility}{Numeric vector of audibilities for the complex spectrum}
}
\value{
\code{data.frame} with columns \code{pitch_midi} and \code{combined_audibility}
}
\description{
Gets the combined spectrum for a sound from its pure and complex audibilities. Represents Equations 11 and 16 in Parncutt & Strasburger (1994)
}
