% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{get_complex_sonorousness}
\alias{get_complex_sonorousness}
\title{Get complex sonorousness}
\usage{
get_complex_sonorousness(complex_tone_audibility, k_c)
}
\arguments{
\item{complex_tone_audibility}{Numeric vector of complex tone audibilities}

\item{k_c}{Parncutt & Strasburger (1994) set this to 0.2 (p. 105)}
}
\value{
Complex sonorousness, a numeric scalar
}
\description{
Computes the complex sonorousness of a sound from its complex tone audibilities.
}
