% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{get_complex_spectrum}
\alias{get_complex_spectrum}
\title{Get complex spectrum}
\usage{
get_complex_spectrum(pitch_midi, pure_tone_audibility,
  midi_params = get_midi_params(), parncutt_params = get_parncutt_params())
}
\arguments{
\item{pitch_midi}{Numeric vector of pure tone MIDI pitches}

\item{pure_tone_audibility}{Corresponding numeric vector of pure tone audibilities}

\item{template_num_harmonics}{The number of harmonics to use in the template}

\item{template_roll_off}{Function to use to compute the level of the nth harmonic in the harmonic template}

\item{k_t}{corresponds to parameter \eqn{k_t} in Equation 10 of Parncutt & Strasburger (1994)}

\item{min_midi}{Minimum MIDI pitch considered in the auditory model}

\item{max_midi}{Maximum MIDI pitch considered in the auditory model}
}
\description{
Gets a complex spectrum from a pure spectrum using a pattern-matching algorithm.
}
