% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{get_multiplicity}
\alias{get_multiplicity}
\title{Represents Equations 14 and 15 from Parncutt & Strasburger (1994)}
\usage{
get_multiplicity(combined_audibility, k_s)
}
\arguments{
\item{combined_audibility}{Numeric vector of combined audibilities as derived in Equation 11 of Parncutt & Strasburger (1994)}
}
\value{
Numeric scalar corresponding to the multiplicity of the sonority
}
\description{
Represents Equations 14 and 15 from Parncutt & Strasburger (1994)
}
