% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_parncutt_params}
\alias{get_parncutt_params}
\title{Get parameters for Parncutt's model}
\usage{
get_parncutt_params(template_num_harmonics = 11, template_roll_off = 1,
  template_stretched_octave = TRUE, k_t = 3, k_p = 0.5, k_c = 0.2,
  k_s = 0.5, min_midi = 0, max_midi = 120)
}
\arguments{
\item{k_t}{Numeric scalar; parameter from Parncutt & Strasburger (1994)}

\item{k_p}{Numeric scalar; parameter from Parncutt & Strasburger (1994)}

\item{k_c}{Numeric scalar; parameter from Parncutt & Strasburger (1994)}

\item{k_s}{Numeric scalar; parameter from Parncutt & Strasburger (1994)}

\item{min_midi}{Numeric scalar; the lowest MIDI pitch considered in the psychoacoustic model}

\item{max_midi}{Numeric scalar; the highest MIDI pitch considered in the psychoacoustic model}
}
\description{
Gets parameters for Parncutt's psychoacoustic model
}
