% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{get_parncutt_pitch_commonality}
\alias{get_parncutt_pitch_commonality}
\title{Get pitch commonality}
\usage{
get_parncutt_pitch_commonality(chord_1, chord_2,
  midi_params = get_midi_params(), parncutt_params = get_parncutt_params(),
  cache_inner = TRUE, cache_outer = FALSE)
}
\arguments{
\item{chord_1}{The first chord to compare}

\item{chord_2}{The second chord to compare}

\item{min_midi}{Numeric scalar; the lowest MIDI pitch considered in the psychoacoustic model}
}
\description{
Gets the pitch commonality between two chords.
}
