% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{get_pitch_distance}
\alias{get_pitch_distance}
\title{Get pitch distance}
\usage{
get_pitch_distance(chord_1, chord_2, min_midi = 0, max_midi = 120)
}
\arguments{
\item{chord_1}{The first chord to compare}

\item{chord_2}{The second chord to compare}

\item{min_midi}{Numeric scalar; the lowest MIDI pitch considered in the psychoacoustic model}
}
\description{
Gets the pitch distance between two chords.
}
