% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{get_pure_sonorousness}
\alias{get_pure_sonorousness}
\title{Get pure sonorousness}
\usage{
get_pure_sonorousness(pure_tone_audibility, k_p)
}
\arguments{
\item{pure_tone_audibility}{Numeric vector of pure tone audibilities}

\item{k_p}{Parncutt & Strasburger (1994) set this to 0.5 (p. 105)}
}
\value{
Pure sonorousness, a numeric scalar
}
\description{
Computes the pure sonorousness of a sound from its pure tone audibilities.
}
