% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{level-class}
\alias{level-class}
\title{Level of a sound.}
\description{
Level of a sound.
}
\section{Slots}{

\describe{
\item{\code{value}}{The value of the level attribute (possibly vectorised for multiple sounds)}
}}

