% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{overall_masking_level}
\alias{overall_masking_level}
\title{Get overall masking level}
\usage{
overall_masking_level(auditory_level, pure_tone_height, k_m = 12)
}
\arguments{
\item{auditory_level}{Numeric vector of auditory levels}

\item{pure_tone_height}{Numeric vector of pure tone heights}
}
\value{
Numeric vector of overall masking levels (dB)
}
\description{
Returns overall masking levels for a set of pure tones that are assumed to be playing simultaneously. Corresponds to Parncutt & Strasburger (1994) Equation 5.
}
