% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parn94.R
\name{parn94}
\alias{parn94}
\alias{parn94.numeric}
\alias{parn94.pi_chord}
\alias{parn94.pi_sparse_spectrum}
\title{Parncutt & Strasburger (1994)}
\usage{
parn94(x, par = parn94_params())

\method{parn94}{numeric}(x, par = parn94_params())

\method{parn94}{pi_chord}(x, par = parn94_params())

\method{parn94}{pi_sparse_spectrum}(x, par = parn94_params())
}
\arguments{
\item{x}{Object to analyse.}

\item{par}{Parameter list as created by \code{\link{parn94_params}()}.}
}
\value{
An list of class \code{parn94}, comprising the following components:
\item{pure_spectrum}{A tibble describing the sonority's pure spectrum.
The pure spectrum is a spectral representation of the input sound
after auditory masking, but before pattern matching.}
\item{pure_spectrum}{A tibble describing the sonority's complex spectrum.
The complex spectrum is created from the pure spectrum through
harmonic template matching.}
\item{combined_spectrum}{A tibble describing the sonority's combined spectrum.
The combined spectrum corresponds to the combination of the
pure and complex spectra.}
\item{par}{A list comprising the parameters used to perform the analysis,
as created by \code{\link{parn94_params}()}.}
}
\description{
This function analyses a sonority using Richard Parncutt's
psychoacoustic model of harmony, as described in
\insertCite{Parncutt1994;textual}{parn94}.
}
\references{
\insertAllCited{}
}
