% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{parncutt_free_field_threshold}
\alias{parncutt_free_field_threshold}
\title{Get free field threshold
Returns the free-field threshold of hearing in quiet for a pure tone of a given frequency. This is the minimum sound level at which a pure tone at that frequency will be heard. Represents Equation 2 in Parncutt & Strasburger (1994).}
\usage{
parncutt_free_field_threshold(freq)
}
\description{
Get free field threshold
Returns the free-field threshold of hearing in quiet for a pure tone of a given frequency. This is the minimum sound level at which a pure tone at that frequency will be heard. Represents Equation 2 in Parncutt & Strasburger (1994).
}
