% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{parncutt_partial_masking_level}
\alias{parncutt_partial_masking_level}
\title{Equation 4 in Parncutt & Strasburger (1994).}
\usage{
parncutt_partial_masking_level(masker, maskee)
}
\description{
Equation 4 in Parncutt & Strasburger (1994).
}
