% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{parncutt_pure_tone_height}
\alias{parncutt_pure_tone_height}
\title{Equation 3 in Parncutt & Strasburger (1994).}
\usage{
parncutt_pure_tone_height(freq)
}
\description{
Equation 3 in Parncutt & Strasburger (1994).
}
