% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch-salience.R
\name{pitch_salience}
\alias{pitch_salience}
\alias{pitch_salience.default}
\alias{pitch_salience.parn94}
\title{Get expanded salience vector}
\usage{
pitch_salience(x, ...)

\method{pitch_salience}{default}(x, ...)

\method{pitch_salience}{parn94}(x)
}
\arguments{
\item{...}{Further arguments to pass to \code{\link{parn94}()}.}

\item{Object}{to analyse.}
}
\description{
Returns a numeric salience vector where each element describes
the salience of a different chromatic pitch.
The first element of this vector corresponds to the
\code{min_midi} argument from \code{\link{parn94_params}},
and the last element corresponds to the \code{max_midi} argument.
}
