% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{pure_tone_height}
\alias{pure_tone_height}
\title{Get pure-tone height}
\usage{
pure_tone_height(kHz)
}
\value{
Numeric vector of corresponding pure-tone heights, with units of equivalent rectangular bandwidths (ERBs)
}
\description{
Returns the pure-tone heights (a.k.a. critical-band rates) of pure tones of given frequencies.
Equation 3 in Parncutt & Strasburger (1994).
#' @param kHz Numeric vector of frequencies in kHz
}
