% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{analyse_sonority}
\alias{analyse_sonority}
\title{Analyse sonority}
\usage{
analyse_sonority(pitch_midi, level_dB = 60, expand_harmonics = TRUE,
  num_harmonics = 11, harmonic_roll_off = function(x) 1/(1 + x), k_t = 3,
  k_p = 0.5, k_c = 0.2, k_s = 0.5, min_midi = 0, max_midi = 120)
}
\arguments{
\item{pitch_midi}{Numeric vector of MIDI pitches in the sonority}

\item{level_dB}{Numeric vector of sound levels (dB) for these pitches; must either be length 1 or of the same length as \code{pitch_midi}}

\item{expand_harmonics}{Boolean scalar; whether or not to expand these pitches to include their implied harmonics}

\item{num_harmonics}{Numeric scalar; if \code{expand_harmonics} is \code{TRUE}, this determines how many harmonics are provided for each pitch, including the fundamental}

\item{harmonic_roll_off}{Function describing the roll off in sound levels as a function of harmonic number}

\item{k_t}{Numeric scalar; parameter from Parncutt & Strasburger (1994)}

\item{k_p}{Numeric scalar; parameter from Parncutt & Strasburger (1994)}

\item{k_c}{Numeric scalar; parameter from Parncutt & Strasburger (1994)}

\item{k_s}{Numeric scalar; parameter from Parncutt & Strasburger (1994)}

\item{min_midi}{Numeric scalar; the lowest MIDI pitch considered in the psychoacoustic model}

\item{max_midi}{Numeric scalar; the highest MIDI pitch considered in the psychoacoustic model}
}
\description{
Analyse a sonority using Parncutt's psychoacoustic models.
}
