% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\name{get_midi_spectrum}
\alias{get_midi_spectrum}
\title{Get MIDI spectrum}
\usage{
get_midi_spectrum(object, keep_inaudible = FALSE, level = 60,
  num_harmonics = 10, roll_off = function(x) 1/(1 + x), min_midi = 0,
  max_midi = 120)
}
\description{
MIDI spectra coerce harmonics to the nearest semitone.
}
