% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parncutt.R
\docType{methods}
\name{initialize,sonority_analysis-method}
\alias{initialize,sonority_analysis-method}
\title{Make new sonority analysis object}
\usage{
\S4method{initialize}{sonority_analysis}(.Object, pitch_midi, level,
  keep_inaudible = FALSE, template_num_harmonics = 11,
  template_roll_off = function(x) 1/(1 + x), stretched_octave = TRUE,
  k_t = 3, k_p = 0.5, k_c = 0.2, k_s = 0.5, min_midi = 0,
  max_midi = 120)
}
\arguments{
\item{k_t}{Parncutt & Strasburger (1994) set this to 3 (p. 104)}

\item{k_p}{Parncutt & Strasburger (1994) set this to 0.5 (p. 105)}

\item{k_c}{Parncutt & Strasburger (1994) set this to 0.2 (p. 105)}

\item{k_s}{Parncutt & Strasburger (1994) set this to 0.5 (p. 106)}
}
\value{
A \code{sonority_analysis} object containing slots describing various aspects of the analysed sonority, including \code{pure_spectrum}, \code{complex_spectrum}, \code{combined_spectrum}, \code{pure_sonorousness}, \code{complex_sonorousness}, and \code{multiplicity}. These features can be accessed using `@`, e.g. `analysis@multiplicity`.
}
\description{
Make new sonority analysis object
}
